整理了searchCourseMock，去除了多余注释
